LayoutTests/inspector-obsolete/http-tests/web-socket-frame-error-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8000/websocket/tests/hybi/does_not_exist' failed: Unexpected response code: 404
Closed.
Tests that WebSocketFrames errors are visible to Web Inspector.

1: Unexpected response code: 404
